# apispec.json
{
    "openapi":"3.0.3",
    "info": {
      "title":"Inventory System RESTful API",
      "description": "API Spec for Inventory System RESTful API",
      "version":"1.0"
    },
    "servers": [
      {"url":"https://localhost:3000/api"}
    ],
    "paths": {
...
        "/auth/profile":{
            "get":{
                "tags": ["Auth API"],
                "security": [{
                    "bearerAuth":[]
                }],
                "description": "Get profile account information",
                "summary": "Profile",
                "responses": {
                    "200":{
                        "description": "Success get profile account information",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type": "object",
                                    "properties": {
                                        "code":{
                                            "type": "number"
                                        },
                                        "status":{
                                            "type": "string"
                                        },
                                        "data":{
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put":{
                "tags": ["Auth API"],
                "security": [{
                    "bearerAuth":[]
                }],
                "description": "Update profile account information",
                "summary": "Update Profile",
                "requestBody":{
                    "description": "at least there must be 1 updated",
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/UpdateAccount"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success update profile account information",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type": "object",
                                    "properties": {
                                        "code":{
                                            "type": "number"
                                        },
                                        "status":{
                                            "type": "string"
                                        },
                                        "data":{
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "schemas": {
            "UserResponse":{
                "type": "object",
                "properties": {
                    "id":{
                        "type": "string"
                    },
                    "name":{
                        "type": "string"
                    },
                    "email":{
                        "type": "string"
                    },
                    "role":{
                        "type": "string"
                    },
                    "status":{
                        "type": "string"
                    },
                    "created_at":{
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at":{
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "UpdateAccount":{
                "type": "object",
                "properties": {
                    "name":{
                        "type": "string"
                    },
                    "email":{
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth":{
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [{
        "bearerAuth":[]
    }]
}